# contracts/runner-api/openapi.yaml
openapi: 3.1.0
info:
  title: Composed MCP Runner API
  version: 0.1.0
paths:
  /runs:
    post:
      summary: Create run pod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_ref, network_policy_profile]
              properties:
                image_ref: { type: string }
                command: { type: array, items: { type: string } }
                args: { type: array, items: { type: string } }
                env_allowlist: { type: object, additionalProperties: { type: string } }
                allowed_tools: { type: array, items: { type: string } }
                downstream_port: { type: integer, minimum: 1, maximum: 65535 }
                cpu: { type: string }
                memory: { type: string }
                timeout_seconds: { type: integer, minimum: 1 }
                network_policy_profile: { type: string, enum: [deny-all, dns-only] }
      responses:
        '201':
          description: Created
        '403':
          description: Policy denied
  /runs/{run_id}:
    get:
      summary: Get run status
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /runs/{run_id}/logs:
    get:
      summary: Get run logs
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /runs/{run_id}/stop:
    post:
      summary: Stop run
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string }
      responses:
        '202': { description: Accepted }
  /runs/{run_id}/tools/{tool_name}:
    post:
      summary: Invoke a downstream tool through runner proxy
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string }
        - in: path
          name: tool_name
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                input:
                  type: object
                  additionalProperties: true
      responses:
        '200': { description: OK }
        '403': { description: Tool not allowed }